chore(tests): split out JS and TS tests
