+++ tests/lib/rules/decorator-position-js.js
@@ -3,7 +3,6 @@
 //------------------------------------------------------------------------------
 
 const parser = require.resolve('babel-eslint');
-const tsParser = require.resolve('@typescript-eslint/parser');
 
 const { stripIndent } = require('common-tags');
 const RuleTester = require('eslint').RuleTester;
@@ -16,7 +15,6 @@ const rule = require('../../../lib/rules/decorator-position');
 //------------------------------------------------------------------------------
 
 const ruleTester = new RuleTester({ parser });
-const tsRuleTester = new RuleTester({ parser: tsParser });
 
 ruleTester.run('JS: decorator-position', rule, {
   valid: [
@@ -318,48 +316,6 @@ ruleTester.run('JS: decorator-position', rule, {
         @foo foo3() {}
       }
     `,
-    },
-  ],
-});
-
-tsRuleTester.run('TS: decorator-position', rule, {
-  valid: [
-    {
-      code: stripIndent`
-        export default class LocaleSwitcher extends Component<IArgs> {
-          @service locale!: LocaleService;
-        }
-      `,
-      options: [{ overrides: { 'prefer-inline': ['@service'] } }],
-    },
-    {
-      code: stripIndent`
-        export default class Foo {
-          @service('addon-name/-private/do-not-use/the-name-of-the-service')
-          declare someObfuscatedPrivateService: Service;
-
-          @service('addon-name/-private/do-not-use/the-name-of-the-service')
-          declare shorterName: Service;
-        }
-      `,
-      options: [{ printWidth: 100 }],
-    },
-  ],
-  invalid: [
-    {
-      code: stripIndent`
-        export default class LocaleSwitcher extends Component<IArgs> {
-          @service
-          locale!: LocaleService;
-        }
-      `,
-      options: [{ overrides: { 'prefer-inline': ['@service'] } }],
-      errors: [{ message: 'Expected @service to be inline.' }],
-      output: stripIndent`
-        export default class LocaleSwitcher extends Component<IArgs> {
-          @service locale!: LocaleService;
-        }
-      `,
     },
   ],
 });
